migrates ray deployments to statefulsets
required to leverage pvc templates for added storage
